Capture logs from terminated containers
If a container has a very short life, sometimes it is captured in the
logs collected by kail and sometimes not. This occurs when the container
exits before kail is able to see the container is alive and collect
logs. This should be a backwards compatible change since the
SinceSeconds value is generally quite small and still applies.